Added statistics

+++ BlitzCache/NullBlitzCacheForTesting.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Runtime.CompilerServices;
 using System.Threading.Tasks;
+using BlitzCacheCore.Statistics;
 
 namespace BlitzCacheCore
 {
@@ -56,7 +57,7 @@ namespace BlitzCacheCore
         public long HitCount => 0;
         public long MissCount => 0;
         public double HitRatio => 0.0;
-        public int CurrentEntryCount => 0;
+        public long CurrentEntryCount => 0;
         public long EvictionCount => 0;
         public int ActiveSemaphoreCount => 0;
         public long TotalOperations => 0;
